feat: 전체 Member의 이름 조회 기능 추가

+++ src/main/java/roomescape/member/ui/MemberRestController.java
@@ -1,9 +1,11 @@
 package roomescape.member.ui;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -12,7 +14,9 @@ import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.domain.AuthRole;
 import roomescape.auth.domain.RequiresRole;
 import roomescape.member.application.MemberService;
+import roomescape.member.domain.Member;
 import roomescape.member.ui.dto.CreateMemberRequest;
+import roomescape.member.ui.dto.MemberNameResponse;
 
 @RestController
 @RequiredArgsConstructor
@@ -32,10 +36,30 @@ public class MemberRestController {
 
     @DeleteMapping("/{id}")
     @RequiresRole(authRoles = {AuthRole.ADMIN, AuthRole.MEMBER})
-    public ResponseEntity<Void> delete(@PathVariable Long id) {
+    public ResponseEntity<Void> delete(
+            @PathVariable final Long id,
+            final Member member
+    ) {
+        if (id == null || member == null) {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST)
+                    .build();
+        }
+        if (!member.getId().equals(id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .build();
+        }
         memberService.delete(id);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT)
                 .build();
+    }
+
+    @GetMapping
+    @RequiresRole(authRoles = {AuthRole.ADMIN})
+    public ResponseEntity<List<MemberNameResponse>> findAll() {
+        final List<MemberNameResponse> memberNameResponses = memberService.findAll();
+
+        return ResponseEntity.status(HttpStatus.OK)
+                .body(memberNameResponses);
     }
 }
